+++ src/minimizer/Minimizer.cpp
@@ -1,14 +1,17 @@
 // Copyright (C) 2018  Philipp Basler and Margarete Mühlleitner
-// SPDX-FileCopyrightText: 2021 Philipp Basler, Margarete Mühlleitner and Jonas Müller
+// SPDX-FileCopyrightText: 2021 Philipp Basler, Margarete Mühlleitner and Jonas
+// Müller
 //
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 #include <BSMPT/minimizer/Minimizer.h>
 #include <BSMPT/models/ClassPotentialOrigin.h> // for Class_Potential_Origin
 #include <BSMPT/models/IncludeAllModels.h>     // for FChoose
-#include <algorithm>                           // for copy, max
-#include <iostream>                            // for operator<<, cout, endl
-#include <math.h>                              // for abs, log10
+#include <BSMPT/utility/Logger.h>
+#include <BSMPT/utility/utility.h>
+#include <algorithm> // for copy, max
+#include <iostream>  // for operator<<, cout, endl
+#include <math.h>    // for abs, log10
 #include <memory>
 #include <random>
 #include <thread>
@@ -194,6 +197,7 @@ Minimize_gen_all(const std::shared_ptr<Class_Potential_Origin> &modelPointer,
 #ifdef NLopt_FOUND
   if (UseMultithreading and thread_NLopt.joinable())
   {
+    Logger::Write(LoggingLevel::MinimizerDetailed, "Waiting for NLopt thread");
     thread_NLopt.join();
   }
 
@@ -201,12 +205,17 @@ Minimize_gen_all(const std::shared_ptr<Class_Potential_Origin> &modelPointer,
   {
     PotValues.push_back(NLOPTResult.PotVal);
     Minima.push_back(NLOPTResult.Minimum);
+    std::stringstream ss;
+    ss << "NLopt candidate at T = " << Temp << " :  " << NLOPTResult.Minimum
+       << " with potential value " << NLOPTResult.PotVal << std::endl;
+    Logger::Write(LoggingLevel::MinimizerDetailed, ss.str());
   }
 #endif
 
 #ifdef libcmaes_FOUND
   if (UseMultithreading and thread_CMAES.joinable())
   {
+    Logger::Write(LoggingLevel::MinimizerDetailed, "Waiting for CMAes Thread");
     thread_CMAES.join();
     auto errC          = LibCMAES.CMAESStatus;
     auto solCMAES      = LibCMAES.result;
@@ -214,11 +223,18 @@ Minimize_gen_all(const std::shared_ptr<Class_Potential_Origin> &modelPointer,
     PotValues.push_back(modelPointer->VEff(solCMAESPotIn, Temp));
     Minima.push_back(solCMAES);
     Check.push_back(errC);
+
+    std::stringstream ss;
+    ss << "CMAES candidate at T = " << Temp << " : " << solCMAES
+       << " with potential value = " << PotValues.at(PotValues.size() - 1)
+       << std::endl;
+    Logger::Write(LoggingLevel::MinimizerDetailed, ss.str());
   }
 #endif
 
   if (UseMultithreading and thread_GSL.joinable())
   {
+    Logger::Write(LoggingLevel::MinimizerDetailed, "Waiting for GSL thread");
     thread_GSL.join();
   }
 
@@ -227,6 +243,12 @@ Minimize_gen_all(const std::shared_ptr<Class_Potential_Origin> &modelPointer,
     solGSLMinPot = modelPointer->MinimizeOrderVEV(solGSLMin);
     PotValues.push_back(modelPointer->VEff(solGSLMinPot, Temp));
     Minima.push_back(solGSLMin);
+
+    std::stringstream ss;
+    ss << "GSL found a minimum at T = " << Temp << ": (" << solGSLMin
+       << ") with Potential value = " << PotValues.at(PotValues.size() - 1)
+       << std::endl;
+    Logger::Write(LoggingLevel::MinimizerDetailed, ss.str());
   }
 
   std::size_t minIndex = 0;