+++ src/models/ClassPotentialRN2HDM.cpp
@@ -498,22 +498,23 @@ void Class_Potential_RN2HDM::write() const
   int posMHCS1 = 0;
   int posA     = 0;
   int posN[3];
-  int countposN  = 0;
-  double testsum = 0;
+  int countposN              = 0;
+  double testsum             = 0;
+  const double ZeroThreshold = 1e-5;
 
   for (std::size_t i = 3; i < NHiggs; i++)
   {
     testsum = std::abs(HiggsRot(i, 0)) + std::abs(HiggsRot(i, 1));
-    if (testsum != 0) posMHCS1 = i;
+    if (testsum > ZeroThreshold) posMHCS1 = i;
     testsum = std::abs(HiggsRot(i, 6)) + std::abs(HiggsRot(i, 7)) +
               std::abs(HiggsRot(i, 8));
-    if (testsum != 0)
+    if (testsum > ZeroThreshold)
     {
       posN[countposN] = i;
       countposN++;
     }
     testsum = std::abs(HiggsRot(i, 4)) + std::abs(HiggsRot(i, 5));
-    if (testsum != 0) posA = i;
+    if (testsum > ZeroThreshold) posA = i;
   }
 
   std::vector<double> HiggsMasses;
@@ -814,32 +815,33 @@ void Class_Potential_RN2HDM::TripleHiggsCouplings()
   int countposN = 0;
   int posG1 = 0, posG2 = 0, posG0 = 0;
 
-  double testsum = 0;
+  double testsum             = 0;
+  const double ZeroThreshold = 1e-5;
 
   for (int i = 0; i < 3; i++)
   {
     testsum = std::abs(HiggsRot(i, 0)) + std::abs(HiggsRot(i, 1));
-    if (testsum != 0) posG1 = i;
+    if (testsum > ZeroThreshold) posG1 = i;
     testsum = std::abs(HiggsRot(i, 2)) + std::abs(HiggsRot(i, 3));
-    if (testsum != 0) posG2 = i;
+    if (testsum > ZeroThreshold) posG2 = i;
     testsum = std::abs(HiggsRot(i, 4)) + std::abs(HiggsRot(i, 5));
-    if (testsum != 0) posG0 = i;
+    if (testsum > ZeroThreshold) posG0 = i;
   }
   for (std::size_t i = 3; i < NHiggs; i++)
   {
     testsum = std::abs(HiggsRot(i, 0)) + std::abs(HiggsRot(i, 1));
-    if (testsum != 0) posMHCS1 = i;
+    if (testsum > ZeroThreshold) posMHCS1 = i;
     testsum = std::abs(HiggsRot(i, 2)) + std::abs(HiggsRot(i, 3));
-    if (testsum != 0) posMHCS2 = i;
+    if (testsum > ZeroThreshold) posMHCS2 = i;
     testsum = std::abs(HiggsRot(i, 6)) + std::abs(HiggsRot(i, 7)) +
               std::abs(HiggsRot(i, 8));
-    if (testsum != 0)
+    if (testsum > ZeroThreshold)
     {
       posN[countposN] = i;
       countposN++;
     }
     testsum = std::abs(HiggsRot(i, 4)) + std::abs(HiggsRot(i, 5));
-    if (testsum != 0) posA = i;
+    if (testsum > ZeroThreshold) posA = i;
   }
 
   std::vector<double> HiggsMasses;